Word Editor ver 1.0
A rough version of word editing. It is a button you click in game to
open an editable text field with all the words. You can then edit them
and after you are done editing you click the button again and it updates
the variable devWordList in the CardDirectory script. It doesn't  write
to the directory yet